+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Next version
 
+# 1.2.0
+
 - Add Blast network name (chain ID 238)
 - Add `chainIdToNetworkName` function to convert a chain IDs to the network names used in Mangrove smart contract repos.